refactoring pipeline.engine
